change c standard from c99 -> c89, no conform yet, just successful compilation

+++ src/h/main.h
@@ -12,25 +12,25 @@ struct hhc_core
 {
     struct /* flag */
     {
-        u64 paused: 1;
-        u64 hud: 1;
-        u64 debug: 1;
-        u64 super_debug: 1;
-        u64 fullscreen: 1;
-        u64 menu_open: 1;
-        u64 fps_cap: 1;
-        u64 parse_target: 1;
-        u64 world_loaded: 1;
-        u64 chunk_buf_dirty: 1;
+        u32 paused: 1;
+        u32 hud: 1;
+        u32 debug: 1;
+        u32 super_debug: 1;
+        u32 fullscreen: 1;
+        u32 menu_open: 1;
+        u32 fps_cap: 1;
+        u32 parse_target: 1;
+        u32 world_loaded: 1;
+        u32 chunk_buf_dirty: 1;
     } flag;
 
     struct /* debug */
     {
-        u64 trans_blocks: 1;
-        u64 chunk_bounds: 1;
-        u64 bounding_boxes: 1;
-        u64 chunk_gizmo: 1;
-        u64 chunk_queue_visualizer: 1;
+        u32 trans_blocks: 1;
+        u32 chunk_bounds: 1;
+        u32 bounding_boxes: 1;
+        u32 chunk_gizmo: 1;
+        u32 chunk_queue_visualizer: 1;
     } debug;
 }; /* hhc_core */
